test(cli): extract blame command tests
